requirements.txt
python-telegram-bot==10.1.0
pyocclient==0.4
SQLAlchemy==1.2.10
psycopg2==2.7.5
